feat(login): Use the card component with the login form
